Refactor code into refactor file
